- Decompose CategoryFilter to talon - Remove duplicate search param utility - Address PR feedback
