+++ packages/venia-ui/lib/RootComponents/Search/search.js
@@ -1,9 +1,1 @@
-import React from 'react';
-
-import SearchPage from '../../components/SearchPage';
-
-const Search = props => {
-    return <SearchPage {...props} />;
-};
-
-export default Search;
+export { default } from '../../components/SearchPage';
