+++ packages/venia-ui/lib/components/SearchPage/categoryFilters.js
@@ -1,7 +1,6 @@
-import React, { useCallback } from 'react';
+import React from 'react';
 import { X as CloseIcon } from 'react-feather';
-import { useQuery } from '@apollo/react-hooks';
-import getQueryParameterValue from '@magento/peregrine/lib/util/getQueryParameterValue';
+import { useCategoryFilters } from '@magento/peregrine/lib/talons/SearchPage/useCategoryFilters';
 
 import { mergeClasses } from '../../classify';
 import Icon from '../../components/Icon';
@@ -9,28 +8,14 @@ import GET_CATEGORY_NAME from '../../queries/getCategoryName.graphql';
 import defaultClasses from './categoryFilters.css';
 
 const CategoryFilters = props => {
-    const { categoryId, executeSearch, history, location } = props;
-    const classes = mergeClasses(defaultClasses, props.classes);
-
-    const handleClearCategoryFilter = useCallback(() => {
-        const inputText = getQueryParameterValue({
-            location,
-            queryParameter: 'query'
-        });
-
-        if (inputText) {
-            executeSearch(inputText, history);
-        }
-    }, [executeSearch, history, location]);
-
-    const { loading, error, data } = useQuery(GET_CATEGORY_NAME, {
-        variables: { id: categoryId }
+    const { categoryId, executeSearch } = props;
+    const talonProps = useCategoryFilters({
+        categoryId,
+        executeSearch,
+        query: GET_CATEGORY_NAME
     });
-
-    let queryResult;
-    if (loading) queryResult = 'Loading...';
-    else if (error) queryResult = null;
-    else queryResult = data.category.name;
+    const { handleClearCategoryFilter, queryResult } = talonProps;
+    const classes = mergeClasses(defaultClasses, props.classes);
 
     return (
         <div className={classes.root}>
